Did work on the following:
Created Client class to manage the code for connecting to server and
peers. (moved it out of the chatWindow class). This should improve the
program's granularity.

Created JUnit tests for the following classes: ChatWindow, Client,
Server, Peer.

Note: Did a lot of work but did not quite complete the proposed goals.
I believe the goal was too broad, so in the future I will try and do
much more narrow commits.